+++ src/features/add-book/schemas/visibilitSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+export const VisibilitySchema = z.object({
+  isPublic: z.boolean(),
+})
