+++ src/features/add-book/steps/index.ts
@@ -1,0 +1,5 @@
+export { BasicInfo } from './BasicInfo'
+export { Rating } from './Rating'
+export { Review } from './Review'
+export { Quote } from './Quote'
+export { Visibility } from './Visibility'
